refactor: move re-useable from regex to custom utils
